pyproject.toml
[project]
name = "plagiarism-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.6.2",
    "fastapi[standard]>=0.134.0",
    "jinja2>=3.1.6",
    "nltk>=3.9.3",
    "passlib[argon2]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.4.0",
    "ruff>=0.15.4",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.47",
    "ty>=0.0.19",
]
